LPD-27899 Move createCategories in order to use it out of blogs
