+++ modules/test/playwright/helpers/createCategories.ts
@@ -3,8 +3,8 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import type {ApiHelpers} from '../../../helpers/ApiHelpers';
-import type {postTaxonomyVocabularyTaxonomyCategoryProps} from '../../../helpers/HeadlessAdminTaxonomyApiHelper';
+import type {ApiHelpers} from './ApiHelpers';
+import type {postTaxonomyVocabularyTaxonomyCategoryProps} from './HeadlessAdminTaxonomyApiHelper';
 
 export type TCategory = Omit<
 	postTaxonomyVocabularyTaxonomyCategoryProps,
